Added HomeFeedActivity.
