+++ app/src/main/AndroidManifest.xml
@@ -19,6 +19,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".HomeFeedActivity"/>
+
     </application>
 
 </manifest>
