+++ app/src/main/java/com/example/android/codeforces/Constants.java
@@ -1,0 +1,6 @@
+package com.example.android.codeforces;
+
+public class Constants {
+    public static final String API_URL = "https://codeforces.com/api/user.rating?handle=";
+    public static final String preferredHandleKey = "preferredHandle";
+}
