+++ app/src/main/res/drawable/circular_button.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item >
+        <shape android:shape="oval">
+            <solid android:color="@color/colorPrimary"/>
+            <size android:width="120dp" android:height="120dp"/>
+        </shape>
+    </item>
+</selector>
