+++ app/src/main/res/drawable/rounded_edittext.xml
@@ -1,0 +1,14 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" android:padding="10dp">
+
+    <solid android:color="#FFFFFF"/>
+
+    <stroke android:width="0.5dp" android:color="#BDBDBD" />
+
+    <corners
+        android:bottomRightRadius="15dp"
+        android:bottomLeftRadius="15dp"
+        android:topLeftRadius="15dp"
+        android:topRightRadius="15dp"/>
+
+</shape>
