+++ app/src/main/res/layout/activity_main.xml
@@ -1,30 +1,51 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    android:background="#FFFFFF">
 
+    <ImageView
+        android:id="@+id/ivCodeForces"
+        android:layout_width="120dp"
+        android:layout_height="120dp"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="64dp"
+        android:src="@drawable/codeforces_logo"/>
 
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/contestsAppeared"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:scrollbars="vertical"
-        android:layout_margin="15dp"/>
-
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
+    <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:src="@drawable/ic_arrow_upward_24dp"
-        app:backgroundTint="#FFFFFF"
-        android:backgroundTint="@color/colorPrimary"
-        android:layout_alignParentEnd="true"
-        android:visibility="gone"
-        android:layout_alignParentBottom="true"
-        android:layout_margin="16dp" />
+        android:layout_below="@id/ivCodeForces"
+        android:layout_centerHorizontal="true"
+        android:text="@string/app_name"
+        android:textStyle="bold"
+        android:textSize="18sp" />
+
+    <android.support.design.widget.TextInputEditText
+        android:id="@+id/etPreferredHandle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:layout_margin="10dp"
+        android:maxLines="1"
+        android:inputType="text"
+        android:padding="12dp"
+        android:imeOptions="actionDone"
+        android:background="@drawable/rounded_edittext"
+        android:shadowColor="#000000"
+        android:hint="@string/enter_preferred_handel"/>
+
+    <ImageView
+        android:id="@+id/ivSubmit"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
+        android:layout_marginTop="8dp"
+        android:layout_centerHorizontal="true"
+        android:layout_below="@id/etPreferredHandle"
+        android:elevation="10dp"
+        android:rotation="90"
+        android:padding="8dp"
+        android:background="@drawable/circular_button"
+        android:src="@drawable/ic_arrow_upward_24dp"/>
 
 </RelativeLayout>
